Add files via upload
When we work with a number of different classes that shares the same methods or properties,
we should not repeat ourselves by writing it in every single class repeatedly.
Alternatively, we can pass the methods or properties on to other classes. This paradigm is known as
DRY which stands for "Don't Repeat Yourself". At the moment we have one or more issues to fix with certain properties or methods, we just need to
work on the parent class.

In inhertance concept, there are two terms known, Parent/Base-class and Child/Sub-class. Child-class inherits the methods or properties owned by the Parent-class.